pyproject.toml
[project]
name = "orimi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.36",
    "django>=5.2.3",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "whitenoise>=6.9.0",
]

[tool.ruff]
line-length = 92
target-version = "py312"
[tool.ruff.lint]
select = ["E", "F", "I"]
[tool.ruff.lint.isort]
known-first-party = ["orimi_shelf"]
